third_party/WebKit/LayoutTests/usb/usb.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/device.mojom.js"></script>
<script src="resources/device_manager.mojom.js"></script>
<script src="resources/permission_provider.mojom.js"></script>
<script src="resources/fake-devices.js"></script>
<script src="resources/usb-helpers.js"></script>
<script>
'use strict';

usb_test(usb => {
  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);

  return navigator.usb.getDevices().then(devices => {
    assert_equals(devices.length, 1);
    usb.assertDeviceInfoEquals(devices[0], usb.fakeDevices[0]);
  });
}, 'getDevices returns devices exposed by the DeviceManager service.');

usb_test(usb => {
  let promise = new Promise((resolve, reject) => {
    navigator.usb.addEventListener('connect', e => {
      usb.assertDeviceInfoEquals(e.device, usb.fakeDevices[0]);
      resolve();
    });
  });

  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);
  return promise;
}, 'onconnect event is trigged by adding a device');

usb_test(usb => {
  let promise = new Promise((resolve, reject) => {
    navigator.usb.addEventListener('disconnect', e => {
      usb.assertDeviceInfoEquals(e.device, usb.fakeDevices[0]);
      resolve();
    });
  });

  usb.mockDeviceManager.addMockDevice(usb.fakeDevices[0]);
  usb.mockDeviceManager.removeMockDevice(usb.fakeDevices[0]);
  return promise;
}, 'ondisconnect event is triggered by removing a device');
</script>
